week-1/requirements.txt
greenlet==2.0.1 ; python_version >= "3.10" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "4.0"
numpy==1.24.1 ; python_version < "4.0" and python_version >= "3.10"
pandas==1.5.2 ; python_version >= "3.10" and python_version < "4.0"
psycopg-binary==3.1.8 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.5 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
pytz==2022.7.1 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==1.4.46 ; python_version >= "3.10" and python_version < "4.0"
